Change return type of getWinnings method to int

+++ lab4/Raceable.java
@@ -2,5 +2,5 @@ package lab4;
 
 public interface Raceable {
     double averageWinnings();
-    double getWinnings();
+    int getWinnings();
 }
